+++ crates/symbolicator-service/src/config.rs
@@ -30,33 +30,6 @@ pub enum LogFormat {
     Json,
 }
 
-#[cfg(feature = "https")]
-#[derive(Clone, Debug, Eq, PartialEq, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub struct HTTPSConfig {
-    pub certificate_path: PathBuf,
-    pub key_path: PathBuf,
-}
-
-#[cfg(feature = "https")]
-impl Default for HTTPSConfig {
-    fn default() -> Self {
-        HTTPSConfig {
-            certificate_path: PathBuf::from("cert.pem"),
-            key_path: PathBuf::from("cert.pem"),
-        }
-    }
-}
-
-/// Controls the HTTPS server setup
-#[cfg(feature = "https")]
-#[derive(Clone, Debug, Default, Deserialize)]
-#[serde(default)]
-pub struct ServerConfig {
-    /// HTTPS configuration
-    pub https: Option<HTTPSConfig>,
-}
-
 /// Controls the logging system.
 #[derive(Clone, Debug, Deserialize)]
 #[serde(default)]
@@ -448,18 +421,11 @@ pub struct Config {
     /// Host and port to bind the HTTP webserver to.
     pub bind: String,
 
-    /// Host and port to bind the HTTPS webserver to.
-    #[cfg(feature = "https")]
-    pub bind_https: Option<String>,
-
     /// Configuration for internal logging.
     pub logging: Logging,
 
     /// Configuration for reporting metrics to a statsd instance.
     pub metrics: Metrics,
-
-    #[cfg(feature = "https")]
-    pub server_config: ServerConfig,
 
     /// DSN to report internal errors to
     pub sentry_dsn: Option<Dsn>,
@@ -621,11 +587,7 @@ impl Default for Config {
         Config {
             cache_dir: default_cache_dir(),
             bind: default_bind(),
-            #[cfg(feature = "https")]
-            bind_https: None,
             logging: Logging::default(),
-            #[cfg(feature = "https")]
-            server_config: ServerConfig::default(),
             metrics: Metrics::default(),
             sentry_dsn: None,
             caches: CacheConfigs::default(),